# nestjs-project/src/videos/dto/complete-upload-response.dto.ts
import { ApiProperty } from '@nestjs/swagger';
import { VideoStatus } from '../videos.constants';

export class CompleteUploadResponseDto {
  @ApiProperty({ format: 'uuid' })
  id: string;

  @ApiProperty({ example: 'dQw4w9WgXcQ' })
  slug: string;

  @ApiProperty({ enum: VideoStatus, example: VideoStatus.PROCESSING })
  status: VideoStatus;
}